requirements.txt
solidpython
viewscad
jupyter
